fixed: compiler warning

+++ xbmc/Application.cpp
@@ -3182,8 +3182,8 @@ bool CApplication::ProcessEventServer(float frameTime)
       newEvent.motion.yrel = 0;
       newEvent.motion.state = 0;
       newEvent.motion.which = 0x10;  // just a different value to distinguish between mouse and event client device.
-      newEvent.motion.x = pos.x;
-      newEvent.motion.y = pos.y;
+      newEvent.motion.x = (uint16_t)pos.x;
+      newEvent.motion.y = (uint16_t)pos.y;
       OnEvent(newEvent);  // had to call this to update g_Mouse position
       return OnAction(CAction(ACTION_MOUSE_MOVE, pos.x, pos.y));
     }
